+++ docs/src/examples/components/Transition/Types/TransitionExample.tsx
@@ -1,0 +1,10 @@
+import * as React from 'react'
+import { Transition, Icon } from '@stardust-ui/react'
+
+const TransitionExample = () => (
+  <Transition animationName="spinner">
+    <Icon name="umbrella" circular />
+  </Transition>
+)
+
+export default TransitionExample