+++ src/components/Form/FormField.tsx
@@ -2,34 +2,64 @@ import * as PropTypes from 'prop-types'
 import * as React from 'react'
 
 import { UIComponent, customPropTypes, childrenExist, createShorthandFactory } from '../../lib'
-import { ComponentVariablesInput, ComponentSlotStyle } from '../../themes/types'
-import {
-  Extendable,
-  ReactChildren,
-  ShorthandValue,
-  ShorthandRenderFunction,
-} from '../../../types/utils'
+import { Extendable, ShorthandValue, ShorthandRenderFunction } from '../../../types/utils'
 import Text from '../Text/Text'
-import { default as Slot } from '../Slot/Slot'
 import Input from '../Input/Input'
+import Slot from '../Slot/Slot'
+import { UIComponentProps, ChildrenComponentProps } from '../../lib/commonPropInterfaces'
+import { commonUIComponentPropTypes, childrenComponentPropTypes } from '../../lib/commonPropTypes'
 
-export interface FormFieldProps {
-  as?: any
-  children?: ReactChildren
-  className?: string
+export interface FormFieldProps extends UIComponentProps<any, any>, ChildrenComponentProps {
+  /** A control for the form field. */
   control?: ShorthandValue
+
+  /** The HTML input id. This will be set on the control element and will be use for linking it with the label for correct accessibility. */
   id?: string
+
+  /** A field can have its label next to instead of above it. */
   inline?: boolean
+
+  /** A label for the form field. */
   label?: ShorthandValue
+
+  /** Text message that will be displayed below the control (can be used for error, warning, success messages). */
   message?: ShorthandValue
+
+  /** The HTML input name. */
   name?: string
+
+  /**
+   * A custom render function for the control slot.
+   *
+   * @param {React.ReactType} Component - The computed component for this slot.
+   * @param {object} props - The computed props for this slot.
+   * @param {ReactNode|ReactNodeArray} children - The computed children for this slot.
+   */
   renderControl?: ShorthandRenderFunction
+
+  /**
+   * A custom render function for the label slot.
+   *
+   * @param {React.ReactType} Component - The computed component for this slot.
+   * @param {object} props - The computed props for this slot.
+   * @param {ReactNode|ReactNodeArray} children - The computed children for this slot.
+   */
   renderLabel?: ShorthandRenderFunction
+
+  /**
+   * A custom render function for the message slot.
+   *
+   * @param {React.ReactType} Component - The computed component for this slot.
+   * @param {object} props - The computed props for this slot.
+   * @param {ReactNode|ReactNodeArray} children - The computed children for this slot.
+   */
   renderMessage?: ShorthandRenderFunction
+
+  /** A field can show that input is mandatory. */
   required?: boolean
-  styles?: ComponentSlotStyle
+
+  /** The HTML input type. */
   type?: string
-  variables?: ComponentVariablesInput
 }
 
 /**
@@ -43,74 +73,19 @@ class FormField extends UIComponent<Extendable<FormFieldProps>, any> {
   static create: Function
 
   public static propTypes = {
-    /** An element type to render as (string or function). */
-    as: customPropTypes.as,
-
-    /**
-     *  FormField content for childrenApi.
-     *  @docSiteIgnore
-     */
-    children: PropTypes.node,
-
-    /** Additional CSS class name(s) to apply.  */
-    className: PropTypes.string,
-
-    /** A control for the form field. */
+    ...commonUIComponentPropTypes,
+    ...childrenComponentPropTypes,
     control: customPropTypes.itemShorthand,
-
-    /** The HTML input id. This will be set on the control element and will be use for linking it with the label for correct accessibility. */
     id: PropTypes.string,
-
-    /** A field can have its label next to instead of above it. */
     inline: PropTypes.bool,
-
-    /** A label for the form field. */
     label: customPropTypes.itemShorthand,
-
-    /** Text message that will be displayed below the control (can be used for error, warning, success messages). */
     message: customPropTypes.itemShorthand,
-
-    /** The HTML input name. */
     name: PropTypes.string,
-
-    /**
-     * A custom render function for the control slot.
-     *
-     * @param {React.ReactType} Component - The computed component for this slot.
-     * @param {object} props - The computed props for this slot.
-     * @param {ReactNode|ReactNodeArray} children - The computed children for this slot.
-     */
     renderControl: PropTypes.func,
-
-    /**
-     * A custom render function for the label slot.
-     *
-     * @param {React.ReactType} Component - The computed component for this slot.
-     * @param {object} props - The computed props for this slot.
-     * @param {ReactNode|ReactNodeArray} children - The computed children for this slot.
-     */
     renderLabel: PropTypes.func,
-
-    /**
-     * A custom render function for the message slot.
-     *
-     * @param {React.ReactType} Component - The computed component for this slot.
-     * @param {object} props - The computed props for this slot.
-     * @param {ReactNode|ReactNodeArray} children - The computed children for this slot.
-     */
     renderMessage: PropTypes.func,
-
-    /** A field can show that input is mandatory. */
     required: PropTypes.bool,
-
-    /** Additional CSS styles to apply to the component instance.  */
-    styles: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
-
-    /** The HTML input type. */
     type: PropTypes.string,
-
-    /** Override for theme site variables to allow modifications of component styling via themes. */
-    variables: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
   }
 
   public static defaultProps = {
@@ -183,6 +158,6 @@ class FormField extends UIComponent<Extendable<FormFieldProps>, any> {
   }
 }
 
-FormField.create = createShorthandFactory(FormField, label => ({ label }))
+FormField.create = createShorthandFactory(FormField, 'label')
 
 export default FormField