+++ src/lib/accessibility/Behaviors/Button/buttonBehavior.ts
@@ -2,9 +2,9 @@ import { Accessibility } from '../../types'
 import * as _ from 'lodash'
 
 /**
- * @description
- *  Adds role='button' if element type is other than 'button'. This allows screen readers to handle the component as a button.
- *  Adds attribute 'aria-disabled=true' based on the property 'disabled'. This can be overriden by providing 'aria-disabled' property directly to the component.
+ * @specification
+ * Adds role='button' if element type is other than 'button'. This allows screen readers to handle the component as a button.
+ * Adds attribute 'aria-disabled=true' based on the property 'disabled'. This can be overriden by providing 'aria-disabled' property directly to the component.
  */
 
 const buttonBehavior: Accessibility = (props: any) => ({