+++ src/lib/accessibility/Behaviors/List/basicListBehavior.ts
@@ -2,8 +2,10 @@ import { Accessibility } from '../../types'
 
 /**
  * @description
- * Adds role='list'.
  * The 'list' role is used to identify an element that creates a list.
+ *
+ * @specification
+ * Adds role='list'.
  */
 
 const basicListBehavior: Accessibility = (props: any) => ({