+++ src/lib/customPropTypes.tsx
@@ -1,6 +1,9 @@
 import * as _ from 'lodash/fp'
 import * as PropTypes from 'prop-types'
 import leven from './leven'
+import { ObjectOf } from '../../types/utils'
+
+type SuggestProps = { score: number; suggestion: string }
 
 const typeOf = x => Object.prototype.toString.call(x)
 
@@ -13,7 +16,7 @@ export const as = (...args) =>
 /**
  * Ensure a prop is a valid DOM node.
  */
-export const domNode = (props, propName) => {
+export const domNode = (props: ObjectOf<any>, propName: string) => {
   // skip if prop is undefined
   if (props[propName] === undefined) return
   // skip if prop is valid
@@ -28,16 +31,16 @@ export const domNode = (props, propName) => {
  * Useful for very large lists of options (e.g. Icon name, Flag name, etc.)
  * @param {string[]} suggestions An array of allowed values.
  */
-export const suggest = suggestions => {
+export const suggest = (suggestions: string[]) => {
   if (!Array.isArray(suggestions)) {
     throw new Error('Invalid argument supplied to suggest, expected an instance of array.')
   }
 
-  const findBestSuggestions = _.memoize(str => {
+  const findBestSuggestions = _.memoize((str: string) => {
     const propValueWords = str.split(' ')
 
     return _.flow(
-      _.map(suggestion => {
+      _.map((suggestion: string) => {
         const suggestionWords = suggestion.split(' ')
 
         const propValueScore = _.flow(
@@ -54,7 +57,7 @@ export const suggest = suggestions => {
 
         return { suggestion, score: propValueScore + suggestionScore }
       }),
-      _.sortBy(['score', 'suggestion']),
+      _.sortBy<SuggestProps>(['score', 'suggestion']),
       _.take(3),
     )(suggestions)
   })
@@ -73,7 +76,7 @@ export const suggest = suggestions => {
     return acc
   }, {})
 
-  return (props, propName, componentName) => {
+  return (props: ObjectOf<any>, propName: string, componentName: string) => {
     const propValue = props[propName]
 
     // skip if prop is undefined or is included in the suggestions
@@ -109,7 +112,11 @@ export const suggest = suggestions => {
  * Disallow other props from being defined with this prop.
  * @param {string[]} disallowedProps An array of props that cannot be used with this prop.
  */
-export const disallow = disallowedProps => (props, propName, componentName) => {
+export const disallow = (disallowedProps: string[]) => (
+  props: ObjectOf<any>,
+  propName: string,
+  componentName: string,
+) => {
   if (!Array.isArray(disallowedProps)) {
     throw new Error(
       [
@@ -148,7 +155,12 @@ export const disallow = disallowedProps => (props, propName, componentName) => {
  * Ensure a prop adherers to multiple prop type validators.
  * @param {function[]} validators An array of propType functions.
  */
-export const every = validators => (props, propName, componentName, ...rest) => {
+export const every = (validators: Function[]) => (
+  props: ObjectOf<any>,
+  propName: string,
+  componentName: string,
+  ...rest: any[]
+) => {
   if (!Array.isArray(validators)) {
     throw new Error(
       [
@@ -178,7 +190,12 @@ export const every = validators => (props, propName, componentName, ...rest) => 
  * Ensure a prop adherers to at least one of the given prop type validators.
  * @param {function[]} validators An array of propType functions.
  */
-export const some = validators => (props, propName, componentName, ...rest) => {
+export const some = (validators: Function[]) => (
+  props: ObjectOf<any>,
+  propName: string,
+  componentName: string,
+  ...rest: any[]
+) => {
   if (!Array.isArray(validators)) {
     throw new Error(
       [
@@ -189,20 +206,20 @@ export const some = validators => (props, propName, componentName, ...rest) => {
   }
 
   const errors = _.compact(
-    _.map(validators, validator => {
+    _.map(validator => {
       if (!_.isFunction(validator)) {
         throw new Error(
           `some() argument "validators" should contain functions, found: ${typeOf(validator)}.`,
         )
       }
       return validator(props, propName, componentName, ...rest)
-    }),
+    }, validators),
   )
 
   // fail only if all validators failed
   if (errors.length === validators.length) {
     const error = new Error('One of these validators must pass:')
-    error.message += `\n${_.map(errors, (err, i) => `[${i + 1}]: ${err.message}`).join('\n')}`
+    error.message += `\n${_.map(err => `- ${err.message}`, errors).join('\n')}`
     return error
   }
 
@@ -214,7 +231,12 @@ export const some = validators => (props, propName, componentName, ...rest) => {
  * @param {object} propsShape An object describing the prop shape.
  * @param {function} validator A propType function.
  */
-export const givenProps = (propsShape, validator) => (props, propName, componentName, ...rest) => {
+export const givenProps = (propsShape: object, validator: Function) => (
+  props: ObjectOf<any>,
+  propName: string,
+  componentName: string,
+  ...rest: any
+) => {
   if (!_.isPlainObject(propsShape)) {
     throw new Error(
       [
@@ -268,7 +290,11 @@ export const givenProps = (propsShape, validator) => (props, propName, component
  * Define prop dependencies by requiring other props.
  * @param {string[]} requiredProps An array of required prop names.
  */
-export const demand = requiredProps => (props, propName, componentName) => {
+export const demand = (requiredProps: string[]) => (
+  props: ObjectOf<any>,
+  propName: string,
+  componentName: string,
+) => {
   if (!Array.isArray(requiredProps)) {
     throw new Error(
       [
@@ -297,7 +323,11 @@ export const demand = requiredProps => (props, propName, componentName) => {
  * Ensure an multiple prop contains a string with only possible values.
  * @param {string[]} possible An array of possible values to prop.
  */
-export const multipleProp = possible => (props, propName, componentName) => {
+export const multipleProp = (possible: string[]) => (
+  props: ObjectOf<string | false>,
+  propName: string,
+  componentName: string,
+) => {
   if (!Array.isArray(possible)) {
     throw new Error(
       [
@@ -365,7 +395,12 @@ export const collectionShorthand = every([disallow(['children']), PropTypes.arra
  * @param {string} help A help message to display with the deprecation warning.
  * @param {function} [validator] A propType function.
  */
-export const deprecate = (help, validator) => (props, propName, componentName, ...args) => {
+export const deprecate = (help: string, validator: Function) => (
+  props: ObjectOf<any>,
+  propName: string,
+  componentName: string,
+  ...args: any[]
+) => {
   if (typeof help !== 'string') {
     throw new Error(
       [
@@ -401,3 +436,17 @@ export const deprecate = (help, validator) => (props, propName, componentName, .
 
   return error
 }
+
+export const animation = PropTypes.oneOfType([
+  PropTypes.shape({
+    name: PropTypes.string.isRequired,
+    delay: PropTypes.string,
+    direction: PropTypes.string,
+    duration: PropTypes.string,
+    fillMode: PropTypes.string,
+    iterationCount: PropTypes.string,
+    playState: PropTypes.string,
+    timingFunction: PropTypes.string,
+  }),
+  PropTypes.string,
+])